webapp/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-storage
#google-cloud-bigquery
google-cloud-logging
# for server:
fastapi
uvicorn[standard]
gunicorn
python-multipart  # For form data processing
starlette         # Core dependency for FastAPI
pydantic          # Data validation
appengine-python-standard>=1.0.0
smart_open
smart_open[gcs]
pyyaml
python-dotenv
./vendor/geoflex-0.0.1-py3-none-any.whl
